mv selectize into module (#145)
Put the four drop down menus (pick_atlas, pick_level1 etc.) in separate shiny module.

The module is called in app_server with 'selection <- select_server("atlas", ...)'

The module needs "language" (as reactive), all the data and config. The module returns which atlas the user pick (stored in selection$atlas()) and the filtered data that can be plotted and tabulated (stored in selection$data()).

The module is tested with shiny::testServer, which is currently only in the development version of the shiny package.